+++ partner/src/epilot/sdkconfiguration.py
@@ -21,9 +21,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '2.1.0'
-    gen_version: str = '2.185.0'
-    user_agent: str = 'speakeasy-sdk/python 2.1.0 2.185.0 1.0.0 epilot-partner'
+    sdk_version: str = '2.1.1'
+    gen_version: str = '2.188.1'
+    user_agent: str = 'speakeasy-sdk/python 2.1.1 2.188.1 1.0.0 epilot-partner'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: